Refactor to new Reactour version
