+++ apps/antalmanac/package.json
@@ -33,6 +33,7 @@
         "@mui/material": "^5.11.7",
         "@packages/antalmanac-types": "*",
         "@react-leaflet/core": "^2.1.0",
+        "@reactour/tour": "^3.6.1",
         "@tanstack/react-query": "^4.24.4",
         "@trpc/client": "^10.30.0",
         "@trpc/server": "^10.30.0",
@@ -62,7 +63,6 @@
         "react-leaflet": "^4.2.1",
         "react-router-dom": "^6.8.1",
         "react-split": "^2.0.14",
-        "reactour": "^1.19.2",
         "recharts": "^2.4.2",
         "superjson": "^1.12.3",
         "websoc-fuzzy-search": "^1.0.1",
