+++ apps/antalmanac/src/App.tsx
@@ -5,10 +5,12 @@ import { SnackbarProvider } from 'notistack';
 import { useEffect } from 'react';
 import ReactGA4 from 'react-ga4';
 
+import { TourProvider } from '@reactour/tour';
 import { undoDelete } from './actions/AppStoreActions';
 import AppQueryProvider from './providers/Query';
 import AppThemeProvider from './providers/Theme';
 import AppThemev5Provider from './providers/Themev5';
+
 
 import Home from './routes/Home';
 import Feedback from './routes/Feedback';
@@ -45,9 +47,11 @@ export default function App() {
         <AppQueryProvider>
             <AppThemeProvider>
                 <AppThemev5Provider>
-                    <SnackbarProvider>
-                        <RouterProvider router={BrowserRouter} />
-                    </SnackbarProvider>
+                    <TourProvider steps={[] /** Will be populated by Tutorial component */}>
+                        <SnackbarProvider>
+                            <RouterProvider router={BrowserRouter} />
+                        </SnackbarProvider>
+                    </TourProvider>
                 </AppThemev5Provider>
             </AppThemeProvider>
         </AppQueryProvider>
